Tag management, splash screen & logo, Contact Screen works on web,
* tag management v1.0 .. pushing to review code in the PR

* changed logic to update sharedWith field of Tag update to avoid circular loop problem

* refactored renderTagGroup to be used across code from common_widget

* add edit screen sends user to tag selection after Expense is created. New expense will NOT show the tags field. Edit expense will show.

* tag management is logically complete. Will review in github PR once tomorrow .. will then run it with Claude to see what it has to say .. will then test :)

* - Tag monetary status is now updated in index.ts when a new expense comes in. Locks are used associated with tagId to ensure update stays consistent
- Attaching/removing expense to a tag flow is streamlined. Each tag is attached/removed one by one and error in one tag operation will not stop overall operation but notify user of the error.
- If an Expense attached to multiple tags is updated, the Expense objects under Tags are updated as batch including the Expense under User. So it will all update or it will all fail.

* handled sending update of tag monetary state to the owner of the expense. Also while intimating friends of updated expense, the check that friend might not have kilvishId & probably exist as a User with phone number is done & kilvishUserId of the friend is now updateD

* Claude suggested some fixes which I applied .. now testing the app manually .. so far so good .. committing & pushing to keep the changes

* put receipt in expense detail a nice primary color border. tags also show as selected on the expense detail screen

* signup screen cosmetic changes - disabled button has grey background .. a lot of changes is because of line changes due to formatting

* tag name update now updates all the screens. The _loadData() call in home_screen was buggy which is also now fixed

* made selectContacts Set mandatory to pass to contact screen

* Tag now have 'sharedWithFriends' field which gets updated when tag is shared with a friend. sharedWithFriends field is now populated in firebase function which looks up each friend, if they have only phone number, lookup kilvish user for them & if they dont find it, create one. Then add their kilvish id to sharedWithFriends field, before sending intimation to them

* firestore rules for deleting expense under tag were incorrect. Fixed that. Alongside that, added logs in various files

* Expense detail screen from tag detail did not have tags information. Fixed now

* Expense edit & delete option will only show to the owner of the expense

* non owner see error message while trying to navigate to edit tag screen from expense detail by tapping the tag

* Edit tag only for owner now implemented

* Tested sharing to a contact & it works. Contact is added to friend of the user & also gets added to sharedWithFriends field of the tag .. the user did not get created but that is server side thingy

* Tag add edit screen had a bug that removed user was not removed from the db .. fixed now

* while creating tag, the tagId should be added to accessibleTagIds of the owner. Did the change

* removed add tag option from tag selection. Modified + floating button on home screen to add expense if user is on expense tab or add tag if user is on tag tab

* home screen -> Expense/Tag -> back to home screen, now populate the newly added Expense/Tag without reloading the page on the top of the list. removed 'Add Expense' button from Tag Detail screen for now

* implemented expense delete & removing them from the callee view without refetching DB

* From Edit Expense screen -> Tag Selection, if no tags are selected, the control does not return back to callee screen . .fixed the logic

* added splash screen & launch icons .. the launch icon can have more padding but that will be dealt later

* local contact permission etc is skipped for kIsWeb

* Checked & fixed PublicInfo search in Contact Search screen

---------

+++ lib/expense_detail_screen.dart
@@ -1,16 +1,63 @@
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:intl/intl.dart';
-import 'package:kilvish/add_edit_expense_screen.dart';
+import 'package:kilvish/expense_add_edit_screen.dart';
 import 'package:kilvish/common_widgets.dart';
+import 'package:kilvish/firestore.dart';
 import 'package:kilvish/models.dart';
+import 'package:kilvish/tag_selection_screen.dart';
 import 'style.dart';
 
-class ExpenseDetailScreen extends StatelessWidget {
+class ExpenseDetailScreen extends StatefulWidget {
   final Expense expense;
 
-  const ExpenseDetailScreen({Key? key, required this.expense})
-    : super(key: key);
+  const ExpenseDetailScreen({Key? key, required this.expense}) : super(key: key);
+
+  @override
+  State<ExpenseDetailScreen> createState() => _ExpenseDetailScreenState();
+}
+
+class _ExpenseDetailScreenState extends State<ExpenseDetailScreen> {
+  late Expense _expense;
+  bool _isExpenseOwner = false;
+
+  @override
+  void initState() {
+    super.initState();
+    _expense = widget.expense;
+    // If tags are empty, fetch them
+    if (_expense.tags.isEmpty) {
+      getExpenseTags(_expense.id).then(
+        (List<Tag>? tags) => {
+          if (tags != null && tags.isNotEmpty) {setState(() => _expense.tags.addAll(tags))},
+        },
+      );
+    }
+    _expense.isExpenseOwner().then((bool isOwner) {
+      if (isOwner == true) setState(() => _isExpenseOwner = true);
+    });
+  }
+
+  Future<void> _openTagSelection() async {
+    if (_isExpenseOwner == false) {
+      if (mounted) showError(context, "Tag editing is only for the owner of the expense");
+      return;
+    }
+
+    print("Calling TagSelectionScreen with ${_expense.id}");
+    final result = await Navigator.push(
+      context,
+      MaterialPageRoute(
+        builder: (context) => TagSelectionScreen(initialSelectedTags: _expense.tags, expenseId: _expense.id),
+      ),
+    );
+
+    if (result != null && result is Set<Tag>) {
+      setState(() {
+        _expense.tags = result;
+      });
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -20,25 +67,23 @@ class ExpenseDetailScreen extends StatelessWidget {
         backgroundColor: primaryColor,
         title: Text(
           'Expense Details',
-          style: TextStyle(
-            color: kWhitecolor,
-            fontSize: titleFontSize,
-            fontWeight: FontWeight.bold,
-          ),
+          style: TextStyle(color: kWhitecolor, fontSize: titleFontSize, fontWeight: FontWeight.bold),
         ),
         leading: IconButton(
           icon: Icon(Icons.arrow_back, color: kWhitecolor),
           onPressed: () => Navigator.pop(context),
         ),
         actions: [
-          IconButton(
-            icon: Icon(Icons.edit, color: kWhitecolor),
-            onPressed: () => _editExpense(context),
-          ),
-          IconButton(
-            icon: Icon(Icons.delete, color: kWhitecolor),
-            onPressed: () => _deleteExpense(context),
-          ),
+          if (_isExpenseOwner == true) ...[
+            IconButton(
+              icon: Icon(Icons.edit, color: kWhitecolor),
+              onPressed: () => _editExpense(context),
+            ),
+            IconButton(
+              icon: Icon(Icons.delete, color: kWhitecolor),
+              onPressed: () => _deleteExpense(context),
+            ),
+          ],
         ],
       ),
       body: SingleChildScrollView(
@@ -53,18 +98,11 @@ class ExpenseDetailScreen extends StatelessWidget {
                 Container(
                   width: 80,
                   height: 80,
-                  decoration: BoxDecoration(
-                    color: primaryColor.withOpacity(0.1),
-                    shape: BoxShape.circle,
-                  ),
+                  decoration: BoxDecoration(color: primaryColor.withOpacity(0.1), shape: BoxShape.circle),
                   child: Center(
                     child: Text(
-                      _getInitial(expense.to),
-                      style: TextStyle(
-                        fontSize: 32,
-                        color: primaryColor,
-                        fontWeight: FontWeight.bold,
-                      ),
+                      _getInitial(_expense.to),
+                      style: TextStyle(fontSize: 32, color: primaryColor, fontWeight: FontWeight.bold),
                     ),
                   ),
                 ),
@@ -73,12 +111,8 @@ class ExpenseDetailScreen extends StatelessWidget {
 
                 // To field
                 Text(
-                  'To ${expense.to}',
-                  style: TextStyle(
-                    fontSize: 20,
-                    color: kTextColor,
-                    fontWeight: FontWeight.w600,
-                  ),
+                  'To ${_expense.to}',
+                  style: TextStyle(fontSize: 20, color: kTextColor, fontWeight: FontWeight.w600),
                   textAlign: TextAlign.center,
                 ),
 
@@ -86,64 +120,43 @@ class ExpenseDetailScreen extends StatelessWidget {
 
                 // Amount (big font)
                 Text(
-                  '₹${expense.amount ?? '0'}',
-                  style: TextStyle(
-                    fontSize: 48,
-                    color: primaryColor,
-                    fontWeight: FontWeight.bold,
-                  ),
+                  '₹${_expense.amount}',
+                  style: TextStyle(fontSize: 48, color: primaryColor, fontWeight: FontWeight.bold),
                 ),
 
                 SizedBox(height: 16),
 
                 // Date and time
-                Text(
-                  _formatDateTime(expense.timeOfTransaction),
-                  style: TextStyle(fontSize: 16, color: kTextMedium),
-                ),
+                Text(_formatDateTime(_expense.timeOfTransaction), style: TextStyle(fontSize: 16, color: kTextMedium)),
 
                 SizedBox(height: 32),
 
                 // Tags
-                if (expense.tags.isNotEmpty) ...[
-                  Text(
-                    'Tags',
-                    style: TextStyle(
-                      fontSize: 14,
-                      color: kTextMedium,
-                      fontWeight: FontWeight.w600,
-                    ),
-                  ),
-                  SizedBox(height: 8),
-                  renderTagGroup(tags: expense.tags),
-                  SizedBox(height: 32),
-                ],
-
+                Text(
+                  'Tags (tap to edit)',
+                  style: TextStyle(fontSize: 14, color: kTextMedium, fontWeight: FontWeight.w600),
+                ),
+                SizedBox(height: 8),
+                GestureDetector(
+                  onTap: _openTagSelection,
+                  child: renderTagGroup(tags: _expense.tags),
+                ),
+                SizedBox(height: 32),
                 // Notes (if any)
-                if (expense.notes != null && expense.notes!.isNotEmpty) ...[
+                if (_expense.notes != null && _expense.notes!.isNotEmpty) ...[
                   Container(
                     width: double.infinity,
                     padding: EdgeInsets.all(16),
-                    decoration: BoxDecoration(
-                      color: tileBackgroundColor,
-                      borderRadius: BorderRadius.circular(8),
-                    ),
+                    decoration: BoxDecoration(color: tileBackgroundColor, borderRadius: BorderRadius.circular(8)),
                     child: Column(
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
                         Text(
                           'Notes',
-                          style: TextStyle(
-                            fontSize: 14,
-                            color: kTextMedium,
-                            fontWeight: FontWeight.w600,
-                          ),
+                          style: TextStyle(fontSize: 14, color: kTextMedium, fontWeight: FontWeight.w600),
                         ),
                         SizedBox(height: 8),
-                        Text(
-                          expense.notes!,
-                          style: TextStyle(fontSize: 16, color: kTextColor),
-                        ),
+                        Text(_expense.notes!, style: TextStyle(fontSize: 16, color: kTextColor)),
                       ],
                     ),
                   ),
@@ -151,31 +164,33 @@ class ExpenseDetailScreen extends StatelessWidget {
                 ],
 
                 // Receipt image (if any)
-                if (expense.receiptUrl != null &&
-                    expense.receiptUrl!.isNotEmpty) ...[
+                if (_expense.receiptUrl != null && _expense.receiptUrl!.isNotEmpty) ...[
                   Text(
                     'Receipt',
-                    style: TextStyle(
-                      fontSize: 14,
-                      color: kTextMedium,
-                      fontWeight: FontWeight.w600,
-                    ),
+                    style: TextStyle(fontSize: 14, color: kTextMedium, fontWeight: FontWeight.w600),
                   ),
                   SizedBox(height: 8),
-                  ClipRRect(
-                    borderRadius: BorderRadius.circular(8),
-                    child: Image.network(
-                      expense.receiptUrl!,
-                      fit: BoxFit.cover,
-                      errorBuilder: (context, error, stackTrace) {
-                        return Container(
-                          height: 200,
-                          color: Colors.grey[300],
-                          child: Center(
-                            child: Icon(Icons.error, color: Colors.red),
-                          ),
-                        );
-                      },
+                  Container(
+                    decoration: BoxDecoration(
+                      border: Border.all(
+                        color: primaryColor, // Border color
+                        width: 2.0, // Border width
+                      ),
+                      borderRadius: BorderRadius.circular(10.0), // Optional: for rounded corners
+                    ),
+                    child: ClipRRect(
+                      borderRadius: BorderRadius.circular(8),
+                      child: Image.network(
+                        _expense.receiptUrl!,
+                        fit: BoxFit.cover,
+                        errorBuilder: (context, error, stackTrace) {
+                          return Container(
+                            height: 200,
+                            color: Colors.grey[300],
+                            child: Center(child: Icon(Icons.error, color: Colors.red)),
+                          );
+                        },
+                      ),
                     ),
                   ),
                 ],
@@ -209,13 +224,7 @@ class ExpenseDetailScreen extends StatelessWidget {
   }
 
   void _editExpense(BuildContext context) {
-    // TODO: Navigate to Add/Edit Expense Screen
-    Navigator.push(
-      context,
-      MaterialPageRoute(
-        builder: (context) => AddEditExpenseScreen(expense: expense),
-      ),
-    );
+    Navigator.push(context, MaterialPageRoute(builder: (context) => AddEditExpenseScreen(expense: _expense)));
   }
 
   void _deleteExpense(BuildContext context) {
@@ -224,22 +233,53 @@ class ExpenseDetailScreen extends StatelessWidget {
       builder: (BuildContext context) {
         return AlertDialog(
           title: Text('Delete Expense', style: TextStyle(color: kTextColor)),
-          content: Text(
-            'Are you sure you want to delete this expense?',
-            style: TextStyle(color: kTextMedium),
-          ),
+          content: Text('Are you sure you want to delete this expense?', style: TextStyle(color: kTextMedium)),
           actions: [
             TextButton(
               onPressed: () => Navigator.pop(context),
               child: Text('Cancel', style: TextStyle(color: kTextMedium)),
             ),
             TextButton(
-              onPressed: () {
-                Navigator.pop(context);
-                Navigator.pop(context); // Go back to previous screen
-                ScaffoldMessenger.of(context).showSnackBar(
-                  SnackBar(content: Text('Delete expense feature coming soon')),
+              onPressed: () async {
+                final navigator = Navigator.of(context, rootNavigator: true);
+
+                Navigator.pop(context); // Close confirmation dialog
+
+                // Show non-dismissible loading dialog
+                showDialog(
+                  context: context,
+                  barrierDismissible: false,
+                  builder: (BuildContext loadingContext) {
+                    return PopScope(
+                      canPop: false,
+                      child: AlertDialog(
+                        content: Row(
+                          children: [
+                            CircularProgressIndicator(color: primaryColor),
+                            SizedBox(width: 20),
+                            Text('Deleting expense...'),
+                          ],
+                        ),
+                      ),
+                    );
+                  },
                 );
+
+                try {
+                  await deleteExpense(widget.expense);
+
+                  // Close loading dialog
+                  if (mounted) navigator.pop();
+                  // Close expense detail screen with result
+                  if (mounted) navigator.pop({'deleted': true, 'expense': _expense});
+                } catch (error, stackTrace) {
+                  print("Error in delete expense $error, $stackTrace");
+                  // Close loading dialog
+                  if (mounted) navigator.pop(context);
+
+                  // Show error
+                  if (mounted) showError(context, "Error deleting expense: $error");
+                }
               },
               child: Text('Delete', style: TextStyle(color: errorcolor)),
             ),
